requirements.txt
flask
requests
psutil
wmi
py-cpuinfo
pywin32
pyinstaller
gunicorn
